remove all but anthropic.  remove unnecessary index files.
