+++ packages/agent/src/core/toolAgent/config.ts
@@ -2,7 +2,7 @@ import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 
-import { createProvider, LLMProvider } from '../llm/index.js';
+import { createProvider, LLMProvider } from '../llm/provider.js';
 import { ToolContext } from '../types';
 
 /**
